Give the homotopy polish route its own module
The production polish path is Gauss-Newton on the overdetermined physical
collocation residual: solver.py reaches polish_legacy_solution, which calls
polish_collocation_least_squares.  The earlier square route -- a homotopy in
alpha from the legacy VMEC endpoint to the strong-force operator, advanced by
pseudo-transient continuation with a pseudo-arclength corrector -- shared a
file with it and made polish_driver.py a 2,434-line module whose two halves
answer different questions.

Move the square route to vmex/core/polish_homotopy.py: 33 definitions, every
one byte-identical to its origin.  Nineteen come from polish_driver.py
(polish_strong_root, the SOLVAX continuation shims, the PTC/bordered
preconditioner helpers, _branch_tangent, _arclength_to_target) and fourteen
from polish.py (StrongModeBlockPreconditioner and its two builders, the
unstructured physical chart, strong_physical_residual, strong_root_rank,
strong_projection_diagnostics and the refresh-policy quartet).

The two routes now share one private helper, _corrected_state, which keeps its
home in polish_driver.py; the two-line lift it used to delegate to is written
out there so the shared surface is that one name plus the certificate helpers
PolishConfig, PolishReport, PolishResult, _normalization_fields and
_failed_certificate_checks.  Nothing else in the driver changed: the only
edits to surviving definitions are the inlined lift and two docstring
cross-references.  vmex.__all__ is untouched at 93 names -- polish_strong_root
was never exported -- and vmex/__init__.py is byte-identical to main.

Homotopy tests move with the code.  tests/test_polish_homotopy.py, which the
manifest already routes to the nightly lane, gains the refresh-policy,
physical-chart, projection-diagnostics, mode-block and already-certified
tests, plus three tests split out of mixed ones in
tests/test_polish_preconditioner.py.  The assertion count across the two
modules is unchanged.
