+++ vmex/core/polish_driver.py
@@ -11,7 +11,6 @@ from __future__ import annotations
 from collections.abc import Iterator
 from contextlib import contextmanager
 from dataclasses import dataclass
-from inspect import signature
 from time import perf_counter
 from typing import Any, Literal, NamedTuple
 
@@ -20,9 +19,8 @@ import functools
 import jax
 import jax.numpy as jnp
 import numpy as np
-from solvax import gmres
 
-from .errors import StrongForceCertificationError, StrongForceContinuationError
+from .errors import StrongForceCertificationError
 from .printing import (
     POLISH_AUTO_DECLINED,
     POLISH_SCREEN_HEADER,
@@ -35,70 +33,22 @@ from .printing import (
     polish_screen_line,
 )
 from .polish import (
-    StrongModeBlockPreconditioner,
     StrongPhysicalChart,
     StrongRootRuntime,
     apply_high_order_correction,
     build_low_order_preconditioner,
-    build_strong_physical_block_preconditioner,
-    build_strong_mode_block_preconditioner,
     make_strong_root_runtime,
     make_strong_structured_chart,
     strong_collocation_residual,
-    strong_physical_residual,
-    strong_root_residual,
 )
 from .strong_force import (
     HighOrderEquilibriumState,
     StrongForceReport,
     certify_strong_force,
-    evaluate_strong_force,
     force_error_measures,
 )
 
 Array = object
-
-
-def _solvax_continuation_api() -> tuple[Any, ...]:
-    """Load the continuation extension supplied by the companion SOLVAX PR."""
-
-    try:
-        from solvax import (
-            ContinuationConfig,
-            PseudoTransientConfig,
-            adaptive_continuation,
-            pseudo_arclength_corrector,
-            pseudo_transient_continuation,
-        )
-    except ImportError as error:
-        raise RuntimeError(
-            "strong-force polishing requires a SOLVAX release containing "
-            "adaptive continuation, pseudo-transient continuation, and "
-            "pseudo-arclength correction (uwplasma/SOLVAX#87)"
-        ) from error
-    return (
-        ContinuationConfig,
-        PseudoTransientConfig,
-        adaptive_continuation,
-        pseudo_arclength_corrector,
-        pseudo_transient_continuation,
-    )
-
-
-def _supports_keyword(function: Any, keyword: str) -> bool:
-    """Return whether an installed SOLVAX callable exposes a new keyword."""
-
-    try:
-        return keyword in signature(function).parameters
-    except (TypeError, ValueError):
-        return False
-
-
-def _residual_evaluations(result: Any) -> int:
-    """Read exact work accounting, with a conservative pre-0.19 fallback."""
-
-    nonlinear_steps = getattr(result, "nonlinear_steps", getattr(result, "steps", 0))
-    return int(getattr(result, "residual_evaluations", nonlinear_steps + 1))
 
 
 #: Wall-clock ceiling ``POLISH = AUTO`` commits to before it declines.
@@ -163,7 +113,8 @@ class PolishConfig:
     ``least_squares_initial_damping``, ``max_nonlinear_iterations``,
     ``linear_restart``, ``linear_max_restarts``, and ``fail_policy``.  The
     remaining knobs configure the square homotopy driver
-    :func:`polish_strong_root`, kept for rank and branch diagnostics.  Every
+    :func:`~vmex.core.polish_homotopy.polish_strong_root`, kept for rank and
+    branch diagnostics.  Every
     range stated below is enforced in ``__post_init__``, which also rejects a
     non-finite value for any of the twelve floating-point controls.
 
@@ -431,7 +382,8 @@ class PolishReport:
     traced and nothing here is recomputed.  The fields from
     ``least_squares_cost`` onwards are produced only by the Gauss--Newton
     route :func:`polish_collocation_least_squares` and stay at their ``None``
-    or ``0`` defaults on the homotopy route :func:`polish_strong_root`.
+    or ``0`` defaults on the homotopy route
+    :func:`~vmex.core.polish_homotopy.polish_strong_root`.
 
     Attributes
     ----------
@@ -806,782 +758,19 @@ def polished_wout_state(
     return sample_high_order_state(native, runtime)
 
 
-@dataclass(frozen=True)
-class _IdentityPreconditioner:
-    """Explicit identity action used to benchmark unpreconditioned JFNK."""
-
-    build_seconds: float = 0.0
-
-    def apply(self, rhs, alpha=1.0, dtau=jnp.inf):
-        del alpha, dtau
-        return rhs
-
-
-def _build_mode_block_preconditioner(
-    runtime: StrongRootRuntime,
-    chart: StrongPhysicalChart | None = None,
-) -> StrongModeBlockPreconditioner:
-    """Backward-compatible private seam for the shared block builder."""
-
-    if chart is None:
-        return build_strong_mode_block_preconditioner(runtime)
-    return build_strong_physical_block_preconditioner(runtime, chart)
-
-
-def _solve_residual(
-    vector: jax.Array,
-    runtime: StrongRootRuntime,
-    alpha: jax.Array,
-    chart: StrongPhysicalChart | None = None,
-) -> jax.Array:
-    """Evaluate either the legacy full chart or the structured physical chart."""
-
-    if chart is None:
-        return strong_root_residual(vector, runtime, alpha)
-    return strong_physical_residual(vector, runtime, chart, alpha)
-
-
-def _full_solve_vector(
-    vector: jax.Array,
-    chart: StrongPhysicalChart | None,
-) -> jax.Array:
-    """Lift physical solve coordinates into the existing full root layout."""
-
-    return jnp.asarray(vector) if chart is None else chart.lift(vector)
-
-
-def _continuation_precondition(
-    rhs: jax.Array,
-    alpha: jax.Array,
-    dtau: jax.Array,
-    runtime: StrongRootRuntime,
-    block_preconditioner: StrongModeBlockPreconditioner,
-    chart: StrongPhysicalChart | None = None,
-) -> jax.Array:
-    """Use the exact legacy inverse early and mode bands near strong force."""
-
-    if isinstance(block_preconditioner, _IdentityPreconditioner):
-        return rhs
-    if chart is not None:
-        return block_preconditioner.apply(rhs, alpha, dtau)
-    return jax.lax.cond(
-        jnp.asarray(alpha) < 0.5,
-        lambda value: _low_inverse(value, runtime),
-        lambda value: block_preconditioner.apply(value, alpha, dtau),
-        rhs,
-    )
-
-
 def _corrected_state(
     vector: jax.Array,
     runtime: StrongRootRuntime,
     chart: StrongPhysicalChart | None = None,
 ):
-    vector = _full_solve_vector(vector, chart)
+    # Lift physical solve coordinates into the full root layout.  This is the
+    # one helper the Gauss--Newton and homotopy routes share, so the lift is
+    # written out here rather than kept as a second shared private name.
+    vector = jnp.asarray(vector) if chart is None else chart.lift(vector)
     correction = runtime.layout.unpack(
         jnp.asarray(runtime.coordinate_scale) * jnp.asarray(vector)
     )
     return apply_high_order_correction(runtime.native, correction)
-
-
-def _minimum_signed_jacobian(
-    vector: jax.Array,
-    runtime: StrongRootRuntime,
-    chart: StrongPhysicalChart | None = None,
-) -> jax.Array:
-    state = _corrected_state(vector, runtime, chart)
-    rr, tt, zz = jnp.meshgrid(
-        jnp.asarray(runtime.radial_nodes),
-        jnp.asarray(runtime.theta),
-        jnp.asarray(runtime.zeta),
-        indexing="ij",
-    )
-    samples = evaluate_strong_force(state, rr, tt, zz)
-    signed = float(state.jacobian_sign) * samples.sqrt_g / jnp.maximum(rr, 1.0e-14)
-    return jnp.min(signed)
-
-
-def _low_inverse(rhs: jax.Array, runtime: StrongRootRuntime) -> jax.Array:
-    """Invert the row-scaled low endpoint in reduced vector coordinates."""
-
-    high_rhs = runtime.layout.unpack(
-        jnp.asarray(rhs) / jnp.asarray(runtime.equation_scale)
-    )
-    low_rhs = runtime.transfer.restrict(high_rhs)
-    solution = runtime.low_preconditioner.solve_scaled(low_rhs)
-    return runtime.layout.pack(runtime.transfer.prolong(solution)) / jnp.asarray(
-        runtime.coordinate_scale
-    )
-
-
-def _solve_low_inverse(
-    rhs: jax.Array,
-    runtime: StrongRootRuntime,
-    chart: StrongPhysicalChart | None = None,
-) -> jax.Array:
-    """Apply the legacy inverse in the active solve-coordinate chart."""
-
-    if chart is None:
-        return _low_inverse(rhs, runtime)
-    full_rhs = jnp.asarray(chart.equation_basis) @ (
-        jnp.asarray(rhs) / jnp.asarray(chart.equation_scale)
-    )
-    full_solution = _low_inverse(full_rhs, runtime)
-    return (
-        jnp.asarray(chart.coordinate_basis).T @ full_solution
-    ) / jnp.asarray(chart.coordinate_scale)
-
-
-def _normalized_low_residual_norm(
-    residual: jax.Array,
-    runtime: StrongRootRuntime,
-    chart: StrongPhysicalChart | None = None,
-) -> jax.Array:
-    """Return the low-endpoint RMS before numerical row equilibration."""
-
-    equation_scale = (
-        runtime.equation_scale if chart is None else chart.equation_scale
-    )
-    size = runtime.layout.size if chart is None else chart.size
-    unscaled = jnp.asarray(residual) / jnp.asarray(equation_scale)
-    return jnp.linalg.norm(unscaled) / np.sqrt(float(size))
-
-
-def _ptc_config(config: PolishConfig, *, residual_scale: float) -> Any:
-    _, PseudoTransientConfig, _, _, _ = _solvax_continuation_api()
-    return PseudoTransientConfig(
-        rtol=config.tolerance,
-        # Couple the roundoff floor to a representative residual norm.  Both
-        # tolerances then transform with a harmless positive equation scaling:
-        # unlike a fixed absolute tolerance this cannot accept an unsolved
-        # stage, while unlike ``atol=0`` it does not chase alpha-zero roundoff.
-        # The independent dimensional certificate remains the final gate.
-        atol=config.tolerance * float(residual_scale),
-        max_steps=config.max_nonlinear_iterations,
-        initial_dt=config.ptc_initial_dtau,
-        max_dt=config.ptc_max_dtau,
-        max_backtracks=config.max_backtracks,
-        linear_restart=config.linear_restart,
-        linear_max_restarts=config.linear_max_restarts,
-    )
-
-
-def _continuation_config(config: PolishConfig) -> Any:
-    ContinuationConfig, _, _, _, _ = _solvax_continuation_api()
-    return ContinuationConfig(
-        target=1.0,
-        initial_step=config.alpha_initial_step,
-        min_step=config.alpha_min_step,
-        max_step=config.alpha_max_step,
-        max_stages=config.max_continuation_stages,
-    )
-
-
-def _branch_tangent(
-    vector: jax.Array,
-    alpha: float,
-    runtime: StrongRootRuntime,
-    config: PolishConfig,
-    previous: tuple[jax.Array, jax.Array] | None,
-    block_preconditioner: StrongModeBlockPreconditioner | None = None,
-    chart: StrongPhysicalChart | None = None,
-) -> tuple[jax.Array, jax.Array]:
-    residual = lambda value: _solve_residual(  # noqa: E731
-        value, runtime, alpha, chart
-    )
-    _, jvp = jax.linearize(residual, vector)
-    parameter_direction = _solve_residual(
-        vector, runtime, 1.0, chart
-    ) - _solve_residual(vector, runtime, 0.0, chart)
-    if previous is None:
-        linear = gmres(
-            jvp,
-            -parameter_direction,
-            precond=(
-                (lambda value: _solve_low_inverse(value, runtime, chart))
-                if block_preconditioner is None
-                else lambda value: block_preconditioner.apply(value, alpha)
-            ),
-            restart=config.linear_restart,
-            rtol=min(1.0e-6, config.tolerance),
-            atol=config.tolerance,
-            max_restarts=config.linear_max_restarts,
-        )
-        linear_x = linear.x
-        linear_alpha = jnp.asarray(1.0, dtype=jnp.asarray(linear.x).dtype)
-    else:
-        previous_x, previous_alpha = previous
-
-        def bordered(value):
-            tangent_x, tangent_alpha = value
-            physical = jax.tree.map(
-                jnp.add,
-                jvp(tangent_x),
-                jax.tree.map(lambda item: item * tangent_alpha, parameter_direction),
-            )
-            normalization = (
-                jnp.vdot(previous_x, tangent_x).real
-                + previous_alpha * tangent_alpha
-            )
-            return physical, normalization
-
-        linear = gmres(
-            bordered,
-            (jnp.zeros_like(vector), jnp.asarray(1.0, dtype=vector.dtype)),
-            precond=lambda rhs: _bordered_preconditioner(
-                runtime, previous, block_preconditioner, chart
-            )((vector, jnp.asarray(alpha)), rhs, jnp.inf),
-            restart=config.linear_restart,
-            rtol=min(1.0e-6, config.tolerance),
-            atol=config.tolerance,
-            max_restarts=config.linear_max_restarts,
-        )
-        linear_x, linear_alpha = linear.x
-    if not bool(linear.converged):
-        raise StrongForceContinuationError(
-            "pseudo-arclength tangent solve did not converge",
-            hint="refine the radial representation or increase the linear budget",
-            alpha=float(alpha),
-            residual_norm=float(linear.residual_norm),
-            linear_iterations=int(linear.iterations),
-        )
-    tangent_x = jnp.asarray(linear_x)
-    tangent_alpha = jnp.asarray(linear_alpha, dtype=tangent_x.dtype)
-    norm = jnp.sqrt(jnp.vdot(tangent_x, tangent_x).real + tangent_alpha**2)
-    tangent_x, tangent_alpha = tangent_x / norm, tangent_alpha / norm
-    if previous is not None:
-        orientation = jnp.vdot(tangent_x, previous[0]).real + tangent_alpha * previous[1]
-        sign = jnp.where(orientation < 0.0, -1.0, 1.0)
-        tangent_x, tangent_alpha = sign * tangent_x, sign * tangent_alpha
-    return tangent_x, tangent_alpha
-
-
-def _bordered_preconditioner(
-    runtime: StrongRootRuntime,
-    tangent: tuple[jax.Array, jax.Array],
-    block_preconditioner: StrongModeBlockPreconditioner | None = None,
-    chart: StrongPhysicalChart | None = None,
-):
-    """Return a low-order block-elimination preconditioner for a bordered root."""
-
-    def apply(state, rhs, dtau):
-        return _apply_bordered_preconditioner(
-            state,
-            rhs,
-            dtau,
-            tangent,
-            runtime,
-            block_preconditioner,
-            chart,
-        )
-
-    return apply
-
-
-def _apply_bordered_preconditioner(
-    state,
-    rhs,
-    dtau,
-    tangent,
-    runtime: StrongRootRuntime,
-    block_preconditioner: StrongModeBlockPreconditioner | None = None,
-    chart: StrongPhysicalChart | None = None,
-):
-    """Apply bordered block elimination with dynamic branch data."""
-
-    vector, alpha = state
-    rhs_x, rhs_alpha = rhs
-    tangent_x, tangent_alpha = tangent
-    parameter_direction = _solve_residual(
-        vector, runtime, 1.0, chart
-    ) - _solve_residual(vector, runtime, 0.0, chart)
-    inverse = (
-        (lambda value: _solve_low_inverse(value, runtime, chart))
-        if block_preconditioner is None
-        else lambda value: block_preconditioner.apply(value, alpha, dtau)
-    )
-    q_rhs = inverse(rhs_x)
-    q_parameter = inverse(parameter_direction)
-    schur = tangent_alpha - jnp.vdot(tangent_x, q_parameter).real
-    tiny = jnp.sqrt(jnp.finfo(jnp.asarray(schur).dtype).eps)
-    safe_schur = jnp.where(
-        jnp.abs(schur) > tiny,
-        schur,
-        jnp.where(schur < 0.0, -tiny, tiny),
-    )
-    delta_alpha = (
-        rhs_alpha - jnp.vdot(tangent_x, q_rhs).real
-    ) / safe_schur
-    return q_rhs - q_parameter * delta_alpha, delta_alpha
-
-
-def _arclength_to_target(
-    vector: jax.Array,
-    alpha: float,
-    runtime: StrongRootRuntime,
-    config: PolishConfig,
-    admissible,
-    block_preconditioner: StrongModeBlockPreconditioner | None,
-    initial_tangent: tuple[jax.Array, jax.Array] | None,
-    chart: StrongPhysicalChart | None = None,
-):
-    _, _, _, pseudo_arclength_corrector, pseudo_transient_continuation = (
-        _solvax_continuation_api()
-    )
-    tangent = (
-        _branch_tangent(
-            vector,
-            alpha,
-            runtime,
-            config,
-            None,
-            block_preconditioner,
-            chart,
-        )
-        if initial_tangent is None
-        else initial_tangent
-    )
-    solve_size = runtime.layout.size if chart is None else chart.size
-    residual_scale = np.sqrt(float(solve_size)) / float(
-        runtime.operator_balance
-    )
-    nonlinear = _ptc_config(config, residual_scale=residual_scale)
-    total_nonlinear = total_linear = total_evaluations = 0
-    arclength_residual = lambda value, parameter: _solve_residual(  # noqa: E731
-        value, runtime, parameter, chart
-    )
-    arclength_admissible = lambda value, parameter: admissible(  # noqa: E731
-        value, parameter
-    )
-    parameterized_precondition = (  # noqa: E731
-        lambda state, rhs, dtau, branch_tangent, predictor: (
-            _apply_bordered_preconditioner(
-                state,
-                rhs,
-                dtau,
-                branch_tangent,
-                runtime,
-                block_preconditioner,
-                chart,
-            )
-        )
-    )
-    for step in range(config.max_arclength_steps):
-        predictor = (
-            vector + config.arclength_step * tangent[0],
-            jnp.asarray(alpha) + config.arclength_step * tangent[1],
-        )
-        if _supports_keyword(pseudo_arclength_corrector, "parameterized_precond"):
-            corrector_kwargs = {
-                "parameterized_precond": parameterized_precondition
-            }
-        elif _supports_keyword(pseudo_arclength_corrector, "precond"):
-            corrector_kwargs = {
-                "precond": _bordered_preconditioner(
-                    runtime, tangent, block_preconditioner, chart
-                )
-            }
-        else:
-            corrector_kwargs = {}
-        corrected = pseudo_arclength_corrector(
-            arclength_residual,
-            predictor,
-            tangent=tangent,
-            predictor=predictor,
-            config=nonlinear,
-            admissible=arclength_admissible,
-            **corrector_kwargs,
-        )
-        total_nonlinear += int(corrected.steps)
-        total_linear += int(corrected.linear_iterations)
-        total_evaluations += _residual_evaluations(corrected)
-        if not bool(corrected.converged) or not bool(corrected.linear_converged):
-            return vector, alpha, step, total_nonlinear, total_linear, total_evaluations
-        previous_alpha = alpha
-        vector, alpha_array = corrected.x
-        alpha = float(alpha_array)
-        if (previous_alpha - 1.0) * (alpha - 1.0) <= 0.0:
-            target = pseudo_transient_continuation(
-                lambda value: _solve_residual(value, runtime, 1.0, chart),
-                vector,
-                precond=(
-                    (
-                        lambda state, rhs, dtau: _solve_low_inverse(
-                            rhs, runtime, chart
-                        )
-                    )
-                    if block_preconditioner is None
-                    else lambda state, rhs, dtau: block_preconditioner.apply(
-                        rhs, 1.0, dtau
-                    )
-                ),
-                admissible=lambda value: admissible(value, 1.0),
-                config=nonlinear,
-            )
-            total_nonlinear += int(target.steps)
-            total_linear += int(target.linear_iterations)
-            total_evaluations += _residual_evaluations(target)
-            if bool(target.converged) and bool(target.linear_converged):
-                return (
-                    target.x,
-                    1.0,
-                    step + 1,
-                    total_nonlinear,
-                    total_linear,
-                    total_evaluations,
-                )
-        tangent = _branch_tangent(
-            vector,
-            alpha,
-            runtime,
-            config,
-            tangent,
-            block_preconditioner,
-            chart,
-        )
-    return (
-        vector,
-        alpha,
-        config.max_arclength_steps,
-        total_nonlinear,
-        total_linear,
-        total_evaluations,
-    )
-
-
-def polish_strong_root(
-    runtime: StrongRootRuntime,
-    *,
-    config: PolishConfig | None = None,
-    initial_certificate: StrongForceReport | None = None,
-    chart: StrongPhysicalChart | None = None,
-) -> PolishResult:
-    """Follow the legacy-connected fixed-boundary branch to strong force.
-
-    The square homotopy route, kept for rank and branch diagnostics; the
-    public polishing path is :func:`polish_collocation_least_squares`.  The
-    residual is the convex interpolation ``(1 - alpha) * low + alpha * strong``
-    between the stored legacy raw-force endpoint and the strong-force
-    endpoint.  The driver solves the ``alpha = 0`` endpoint with
-    pseudo-transient continuation (skipping it when the endpoint residual is
-    already at roundoff before row equilibration), runs SOLVAX adaptive
-    continuation to ``alpha = 1``, and, if that stalls and
-    ``use_pseudo_arclength`` is set, retries with bordered pseudo-arclength
-    steps that can pass a fold in ``alpha``.  Every trial state must keep a
-    signed Jacobian above the configured margin, so the branch cannot walk
-    through a surface overlap.  Reaching ``alpha = 1`` is not acceptance: the
-    corrected state is then handed to the independent certificate.
-
-    Parameters
-    ----------
-    runtime:
-        Prepared strong-root runtime from
-        :func:`~vmex.core.polish.make_strong_root_runtime`.  It fixes the
-        collocation grid, the constrained root layout, the row and column
-        equilibration, the transfer to the legacy packing, and the low-order
-        preconditioner.  ``runtime.native`` is the state being corrected.
-    config:
-        Controls; ``None`` uses ``PolishConfig()``.  This route reads the
-        continuation, PTC, arclength, preconditioner, and Jacobian-margin
-        knobs that the Gauss--Newton route ignores.
-    initial_certificate:
-        Certificate of ``runtime.native``, if the caller already has one.
-        ``None`` evaluates it here.  When it already meets
-        ``certificate_tolerance`` the driver returns immediately with
-        ``termination_reason="already-certified"`` and a zero correction.
-    chart:
-        Optional gauge-free physical chart.  ``None`` solves in the full
-        constrained root layout with ``strong_root_residual``; passing a chart
-        solves the projected square system in ``chart.size`` coordinates
-        instead, via ``strong_physical_residual``.
-
-    Returns
-    -------
-    A :class:`PolishResult`.  Its ``context`` is always ``None`` — this route
-    does not build the least-squares stationarity context that the implicit
-    derivative entry points require.
-
-    Raises
-    ------
-    StrongForceContinuationError
-        If the branch does not reach ``alpha = 1`` and ``fail_policy`` is
-        ``"raise"``.
-    StrongForceCertificationError
-        If ``alpha = 1`` is reached but the independent certificate exceeds
-        ``certificate_tolerance`` and ``fail_policy`` is ``"raise"``.
-    """
-
-    config = PolishConfig() if config is None else config
-    started = perf_counter()
-    initial_certificate = (
-        certify_strong_force(runtime.native)
-        if initial_certificate is None
-        else initial_certificate
-    )
-    solve_size = runtime.layout.size if chart is None else chart.size
-    zero = jnp.zeros(
-        (solve_size,), dtype=jnp.asarray(runtime.native.R_cos).dtype
-    )
-    full_zero = jnp.zeros((runtime.layout.size,), dtype=zero.dtype)
-    if not _failed_certificate_checks(initial_certificate, config):
-        report = PolishReport(
-            converged=True,
-            termination_reason="already-certified",
-            final_alpha=1.0,
-            initial_normalized_l2=float(initial_certificate.normalized_l2),
-            final_normalized_l2=float(initial_certificate.normalized_l2),
-            continuation_accepted=0,
-            continuation_rejected=0,
-            nonlinear_iterations=0,
-            linear_iterations=0,
-            residual_evaluations=0,
-            arclength_steps=0,
-            minimum_signed_jacobian=float(initial_certificate.minimum_signed_jacobian),
-            factor_build_seconds=runtime.low_preconditioner.factor_build_seconds,
-            solve_seconds=perf_counter() - started,
-            radial_refinement_tolerance=config.radial_refinement_tolerance,
-            **_normalization_fields(initial_certificate),
-        )
-        return PolishResult(runtime.native, initial_certificate, report, full_zero)
-    _, _, adaptive_continuation, _, pseudo_transient_continuation = (
-        _solvax_continuation_api()
-    )
-    initial_margin = float(_minimum_signed_jacobian(zero, runtime, chart))
-    if config.preconditioner == "mode-block":
-        block_preconditioner = _build_mode_block_preconditioner(runtime, chart)
-    elif config.preconditioner == "none":
-        block_preconditioner = _IdentityPreconditioner()
-    else:
-        block_preconditioner = None
-    factor_build_seconds = (
-        runtime.low_preconditioner.factor_build_seconds
-        + (0.0 if block_preconditioner is None else block_preconditioner.build_seconds)
-    )
-    margin_floor = max(
-        config.minimum_jacobian_floor,
-        config.minimum_jacobian_ratio * initial_margin,
-    )
-
-    def admissible(vector, alpha):
-        del alpha
-        residual = _solve_residual(vector, runtime, 1.0, chart)
-        return (
-            jnp.all(jnp.isfinite(vector))
-            & jnp.all(jnp.isfinite(residual))
-            & (_minimum_signed_jacobian(vector, runtime, chart) >= margin_floor)
-        )
-
-    residual_scale = np.sqrt(float(solve_size)) / float(
-        runtime.operator_balance
-    )
-    nonlinear = _ptc_config(config, residual_scale=residual_scale)
-    precondition = (  # noqa: E731
-        (lambda state, rhs, dtau: rhs)
-        if config.preconditioner == "none"
-        else lambda state, rhs, dtau: _solve_low_inverse(rhs, runtime, chart)
-    )
-    # The low homotopy endpoint subtracts the stored legacy defect, so zero is
-    # its mathematical root.  Roundoff from restrict/project/prolong may leave
-    # a tiny row-equilibrated remainder.  Check that remainder before row
-    # scaling and avoid asking PTC to reduce it below floating-point noise.
-    # A genuinely inconsistent endpoint still takes the globalized solve.
-    endpoint_residual = _solve_residual(zero, runtime, 0.0, chart)
-    endpoint_solved = float(
-        _normalized_low_residual_norm(endpoint_residual, runtime, chart)
-    ) <= config.tolerance
-    if endpoint_solved:
-        vector = zero
-        nonlinear_iterations = 0
-        linear_iterations = 0
-        residual_evaluations = 1
-        converged = True
-    else:
-        endpoint = pseudo_transient_continuation(
-            lambda vector: _solve_residual(vector, runtime, 0.0, chart),
-            zero,
-            precond=precondition,
-            admissible=lambda vector: admissible(vector, 0.0),
-            config=nonlinear,
-        )
-        vector = endpoint.x
-        nonlinear_iterations = int(endpoint.steps)
-        linear_iterations = int(endpoint.linear_iterations)
-        residual_evaluations = 1 + _residual_evaluations(endpoint)
-        converged = bool(endpoint.converged) and bool(endpoint.linear_converged)
-    steps: tuple[Any, ...] = ()
-    arclength_steps = 0
-    alpha = 0.0
-    reason = "alpha-zero-failed"
-    if converged:
-        accepted_states: list[tuple[jax.Array, float]] = [(vector, 0.0)]
-
-        def record_accepted_state(candidate, parameter, solution):
-            del solution
-            accepted_states.append((candidate, float(parameter)))
-            return True
-
-        # Kwargs for adaptive_continuation: the two spellings carry
-        # different callable arities, so the annotation is deliberately Any.
-        continuation_preconditioners: dict[str, Any] = (
-            {"precond": precondition}
-            if block_preconditioner is None
-            or not _supports_keyword(
-                adaptive_continuation, "parameterized_precond"
-            )
-            else {
-                "parameterized_precond": (
-                    lambda state, rhs, dtau, parameter: _continuation_precondition(
-                        rhs,
-                        parameter,
-                        dtau,
-                        runtime,
-                        block_preconditioner,
-                        chart,
-                    )
-                )
-            }
-        )
-        continuation = adaptive_continuation(
-            lambda value, parameter: _solve_residual(
-                value, runtime, parameter, chart
-            ),
-            vector,
-            alpha0=0.0,
-            nonlinear_config=nonlinear,
-            continuation_config=_continuation_config(config),
-            admissible=admissible,
-            accept_stage=record_accepted_state,
-            **continuation_preconditioners,
-        )
-        steps = continuation.steps
-        vector, alpha = continuation.x, continuation.alpha
-        nonlinear_iterations += sum(stage.nonlinear_steps for stage in steps)
-        linear_iterations += sum(stage.linear_iterations for stage in steps)
-        residual_evaluations += sum(_residual_evaluations(stage) for stage in steps)
-        converged = continuation.converged
-        reason = "strong-root" if converged else "continuation-stalled"
-        if not converged and config.use_pseudo_arclength:
-            initial_tangent = None
-            if len(accepted_states) >= 2:
-                previous_vector, previous_alpha = accepted_states[-2]
-                delta_vector = vector - previous_vector
-                delta_alpha = jnp.asarray(alpha - previous_alpha)
-                tangent_norm = jnp.sqrt(
-                    jnp.vdot(delta_vector, delta_vector).real + delta_alpha**2
-                )
-                initial_tangent = (
-                    delta_vector / tangent_norm,
-                    delta_alpha / tangent_norm,
-                )
-            try:
-                (
-                    vector,
-                    alpha,
-                    arclength_steps,
-                    arc_nonlinear,
-                    arc_linear,
-                    arc_evaluations,
-                ) = _arclength_to_target(
-                    vector,
-                    alpha,
-                    runtime,
-                    config,
-                    admissible,
-                    block_preconditioner,
-                    initial_tangent,
-                    chart,
-                )
-            except StrongForceContinuationError:
-                reason = "pseudo-arclength-tangent-failed"
-            else:
-                nonlinear_iterations += arc_nonlinear
-                linear_iterations += arc_linear
-                residual_evaluations += arc_evaluations
-                converged = alpha == 1.0
-                reason = (
-                    "pseudo-arclength" if converged else "pseudo-arclength-stalled"
-                )
-
-    accepted = sum(stage.accepted for stage in steps)
-    rejected = len(steps) - accepted
-    if not converged:
-        report = PolishReport(
-            converged=False,
-            termination_reason=reason,
-            final_alpha=float(alpha),
-            initial_normalized_l2=float(initial_certificate.normalized_l2),
-            final_normalized_l2=float(initial_certificate.normalized_l2),
-            continuation_accepted=accepted,
-            continuation_rejected=rejected,
-            nonlinear_iterations=nonlinear_iterations,
-            linear_iterations=linear_iterations,
-            residual_evaluations=residual_evaluations,
-            arclength_steps=arclength_steps,
-            minimum_signed_jacobian=float(
-                _minimum_signed_jacobian(vector, runtime, chart)
-            ),
-            factor_build_seconds=factor_build_seconds,
-            solve_seconds=perf_counter() - started,
-            **_normalization_fields(initial_certificate),
-        )
-        if config.fail_policy == "raise":
-            raise StrongForceContinuationError(
-                "strong-force continuation did not reach alpha=1",
-                hint="inspect the continuation report and refine the radial representation",
-                alpha=float(alpha),
-                residual_norm=float(
-                    jnp.linalg.norm(
-                        _solve_residual(vector, runtime, alpha, chart)
-                    )
-                ),
-                nonlinear_iterations=nonlinear_iterations,
-                linear_iterations=linear_iterations,
-                accepted_stages=accepted,
-                rejected_stages=rejected,
-            )
-        return PolishResult(runtime.native, initial_certificate, report, full_zero)
-
-    state = _corrected_state(vector, runtime, chart)
-    certificate = certify_strong_force(state)
-    failed_checks = _failed_certificate_checks(certificate, config)
-    certified = not failed_checks
-    report = PolishReport(
-        converged=certified,
-        termination_reason="certified" if certified else "certification-failed",
-        final_alpha=1.0,
-        initial_normalized_l2=float(initial_certificate.normalized_l2),
-        final_normalized_l2=float(certificate.normalized_l2),
-        continuation_accepted=accepted,
-        continuation_rejected=rejected,
-        nonlinear_iterations=nonlinear_iterations,
-        linear_iterations=linear_iterations,
-        residual_evaluations=residual_evaluations,
-        arclength_steps=arclength_steps,
-        minimum_signed_jacobian=float(certificate.minimum_signed_jacobian),
-        factor_build_seconds=factor_build_seconds,
-        solve_seconds=perf_counter() - started,
-        radial_refinement_tolerance=config.radial_refinement_tolerance,
-        **_normalization_fields(initial_certificate, certificate),
-    )
-    if not certified and config.fail_policy == "raise":
-        raise StrongForceCertificationError(
-            "strong root failed the independent force certificate: "
-            + "; ".join(failed_checks),
-            hint="increase radial degree/resolution and retry once",
-            solver_converged=True,
-            normalized_l2=float(certificate.normalized_l2),
-            tolerance=config.certificate_tolerance,
-            radial_refinement=float(certificate.radial_refinement_difference),
-            radial_refinement_tolerance=config.radial_refinement_tolerance,
-        )
-    if not certified:
-        return PolishResult(runtime.native, initial_certificate, report, full_zero)
-    return PolishResult(
-        state,
-        certificate,
-        report,
-        _full_solve_vector(vector, chart),
-    )
 
 
 #: Shortest gap between two live polish heartbeats.  The callbacks that feed
@@ -2430,5 +1619,4 @@ __all__ = [
     "polished_compatibility_state",
     "polished_wout_ns",
     "polished_wout_state",
-    "polish_strong_root",
 ]
